refactor: rename list itme
